requirements.txt
alembic==1.5.1
arrow==1.2.0
asgiref==3.3.1
awesome-slugify==1.6.5
backcall==0.2.0
backports.cached-property==1.0.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
binaryornot==0.4.4
bleach==3.2.3
blessings==1.7
blinker==1.4
bpython==0.21
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==35.0.0
curtsies==0.3.7
cwcwidth==0.1.4
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.1.3
django-allauth==0.45.0
django-bootstrap3==15.0.0
django-braces==1.15.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.10.0
django-crispy-forms==1.12.0
django-debug-toolbar==1.4
django-filer==2.0.2
django-filter==21.1
django-formtools==2.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-mobile==0.7.0
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-mptt-tree-editor==0.1.2
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-tastypie==0.14.3
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.4
dnspython==2.0.0
dominate==2.6.0
drf-yasg==1.20.0
easy-thumbnails==2.7.1
email-validator==1.1.2
Faker==5.7.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-PageDown==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
httpie==2.4.0
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
ipython==7.28.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
jinja2-time==0.2.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mock==4.0.3
oauthlib==3.1.1
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
Pillow==8.0.1
poyo==0.5.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
pudb==2021.1
pycparser==2.20
Pygments==2.8.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
python-memcached==1.59
python-mimeparse==1.6.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2020.4
pyxdg==0.27
PyYAML==6.0
regex==2021.9.30
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.15.0
South==1.0.2
SQLAlchemy==1.3.20
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.1.0
typing-extensions==3.7.4.3
tzdata==2021.1
tzlocal==3.0
Unidecode==0.4.21
uritemplate==4.1.1
urllib3==1.26.4
urwid==2.1.2
visitor==0.1.3
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.3.0
WTForms==2.3.3
xlwt==1.3.0
zipp==3.4.0
